default 50ms loop, 2.5s max len

+++ golden-harp-adapter/golden-harp-adapter.ino
@@ -58,7 +58,7 @@
 
 // cached loop_time value so we're not reading EEPROM every time around the loop
 unsigned long loop_time_ms;
-unsigned long max_note_length_ms;
+unsigned short max_note_length_ms;
 
 
 void setup()
